refactor: view 수정
- 도메인 1대1 대응 view 삭제
- input/output view 추가
